requirements.txt
openai>=1.10.0
python-dotenv>=1.0.0
pandas>=2.0.0
PyMuPDF>=1.22.0
Pillow>=10.0.0